Problem 8: Change billing schedule
When changing a policy's billing schedule, recalculate its invoices beginning
on the original effective_date, not from the point of the most recent invoice.
In all happy paths, the math works out fine.

With the particular scenario given by the README, the math doesn't quite come
out correctly, as we're dealing with Integer columns. I figured that wasn't
what you wanted me to focus on, and solved the larger problem at hand.