Problem 13: Knockout validations
To keep from executing extraneous Javascript, extract how javascript files may
be linked within the application, under the layout.html template as a new block
that may be extended by other templates.

After trying, and failing, with knockout-validation.js, handroll a simple
validation requiring a policy_number be passed into the knockout_search form.
No date validation is required, as the implementation defaults date to current
day. If a request is submitted to the server successfully, but it responds with
a 404 (cannot find the policy), a second validation warning is presented to the
user.

Both validation errors disable the submission of the form until the user
corrects the problem.

I would have liked to prevent the form from submitting at all with no given
data (what I assumed would be your first test of my code). I was unable to
resolve one of two issues: either Knockout-Validation's APIs are too shallow
for how I was attempting to solve the problem, or I haven't quite mastered how
to search for documentation on the library.

If I used the validation library, executing the validations and checking errors
worked flawlessly. The only issue was that I couldn't prevent them from running
when I wanted to (initial pageload). If I rolled my own validation, with no
help from the library, I lost the ability to force a validation (after pageload,
before any of the fields are modified, just before submitting the form). If I
tried to blend the two, my head started spinning and I found myself off-track.

I'd love to hear of your expected solution to this problem, as I'm sure I was
simply making it more difficult than it needed be.