Bonus: Wrap lines at 80 characters
My additions to the code had neglected line wrap suggestions, as the priority
was to get everything up and running.

Are backslashes before newlines acceptable? I saw them a couple places in the
project initially. Not sure if they are common, or if they should be treated
as a last resort to format source code.